+++ src/hash_table.rs
@@ -361,6 +361,46 @@ where
         Ok(post_processor(None))
     }
 
+    /// Retains entries that satisfy the specified predicate.
+    fn retain_entries<F: FnMut(&K, &mut V) -> bool>(&self, mut pred: F) -> (usize, usize) {
+        let barrier = Barrier::new();
+        let mut num_retained: usize = 0;
+        let mut num_removed: usize = 0;
+        let mut current_array_ptr = self.bucket_array().load(Acquire, &barrier);
+        while let Some(current_array) = current_array_ptr.as_ref() {
+            self.clear_old_array(current_array, &barrier);
+            for index in 0..current_array.num_buckets() {
+                let bucket = current_array.bucket_mut(index);
+                if let Some(mut locker) = Locker::lock(bucket, &barrier) {
+                    let data_block_mut = current_array.data_block_mut(index);
+                    let mut entry_ptr = EntryPtr::new(&barrier);
+                    while entry_ptr.next(&locker, &barrier) {
+                        let (k, v) = entry_ptr.get_mut(data_block_mut, &mut locker);
+                        if pred(k, v) {
+                            num_retained = num_retained.saturating_add(1);
+                        } else {
+                            locker.erase(data_block_mut, &mut entry_ptr);
+                            num_removed = num_removed.saturating_add(1);
+                        }
+                    }
+                }
+            }
+
+            let new_current_array_ptr = self.bucket_array().load(Acquire, &barrier);
+            if current_array_ptr.without_tag() == new_current_array_ptr.without_tag() {
+                break;
+            }
+            num_retained = 0;
+            current_array_ptr = new_current_array_ptr;
+        }
+
+        if num_removed >= num_retained {
+            self.try_resize(0, &barrier);
+        }
+
+        (num_retained, num_removed)
+    }
+
     /// Acquires a [`Locker`] and [`EntryPtr`] corresponding to the key.
     ///
     /// It returns an error if locking failed, or returns an [`EntryPtr`] if the key exists,