+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+1.6.2
+
+* API update: add `HashIndex::{retain, retain_async}`.
+
 1.6.1
 
 * API update: add a mutable `Bag` iterator.